rule-finder/backend/requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
chromadb-client
requests
onnxruntime
tokenizers